requirements.txt
Django==1.11
django-ckeditor==5.2.2
Pillow==4.1.1